% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{sdcTarget-package}
\alias{sdcTarget-package}
\title{The sdcTarget Package}
\description{
Statistical Disclosure Control Substitution Matrix Calculator.

Provides functions to calculate target matrices for
  statistical disclosure control.
}
\details{
\tabular{ll}{
Package: \tab sdcTarget\cr
Type: \tab Package\cr
Version: \tab 0.9-11\cr
Date: \tab 2014-11-03\cr
Depends: \tab R (>= 2.10.0), methods, magic, foreach, parallel, doParallel, tuple (>= 0.4-02)\cr
Encoding: \tab UTF-8\cr
License: \tab CC BY-NC 4.0\cr
LazyLoad: \tab no\cr
URL: \tab http://statistics.lazaridis.eu\cr
}

Classes and methods to calculate and evaluate target matrices for statistical disclosure control.

Parallel processing is now implemented by default
  to speed up large searches. The number of processors
  must be specified. If not specified, will default
  to linear processing.
}
\author{
Emmanuel Lazaridis
}
\references{
\href{http://www.statslife.org.uk/events/annual-conference/conference-blog/1752-data-linkage-and-privacy}{Summary of talk at RSS 2014 Conference}
}
\keyword{package}

