% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,sdcHashClass-method}
\alias{show,sdcHashClass-method}
\alias{summary,sdcHashClass-method}
\title{Show Method For \code{\linkS4class{sdcHashClass}} Objects}
\usage{
\S4method{show}{sdcHashClass}(object)

\S4method{summary}{sdcHashClass}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{sdcHashClass}} object.}
}
\description{
Returns an abbreviated summary of a \code{\linkS4class{sdcHashClass}} object.

The summary method runs and returns the output of
  some calculations on an \code{\linkS4class{sdcHashClass}}
  object, together with an abbreviated summary of the
  object itself.
}
\details{
As described in the documentation for \code{\link[methods]{show}},
  the default for printing out S4 objects is a call to
  a \code{\link[methods]{show}} method, whereas
  \code{\link[base]{print}} is the default for S3 objects.  This gives
  the possibility to have an alternative way to print the contents
  of a user-defined S4 class object.
}
\section{Methods (by class)}{
\itemize{
\item \code{sdcHashClass}: 

\item \code{sdcHashClass}: 
}}
\examples{
show(new("sdcHashClass"))
summary(new("sdcHashClass"))
my.X <- data.frame(matrix(ifelse(runif(500)>.5, TRUE, FALSE), ncol = 5))
my.hc <- new("sdcHashClass", X = my.X)
summary(my.hc)
}

