% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[,sdcHashClass-method}
\alias{[,sdcHashClass-method}
\title{Get a Hash List Subset}
\usage{
\S4method{[}{sdcHashClass}(x, i)
}
\arguments{
\item{x}{An \code{\linkS4class{sdcHashClass}} object.}

\item{i}{An index of the hash.}
}
\description{
Returns a subset of a \code{\linkS4class{sdcHashClass}} object.
}
\details{
Does what it says on the package.
}
\examples{
my.X <- data.frame(matrix(ifelse(runif(5000)>.5, TRUE, FALSE), ncol = 50))
my.hc <- new("sdcHashClass", X = my.X)
my.hc[2:3]
}
\seealso{
\code{\linkS4class{sdcHashClass}},
}

